Rename URL module
  - Rename erddapy.core.url_handling to erddapy.core.url
  - Refactor imports
